Modules language and prefecture order
All Japanese modules for shipping , payment and order total will install in admin current language, Japanese or English. With this automatized, file edit to choose language should not be necessary anymore.
Pull down menus for Japanese zone will show prefectures (states) when in Japanese in standard administrative order, which is North to South.